requirement.txt
pytest
pytest-html
pytest-xdist
numpy
pyecharts
torch
onnx
onnx-simplifier
bidict
toml
hnn